requirements.txt
django 
djangorestframework 
psycopg2
djangorestframework-simplejwt
pillow
django-cors-headers
python-decouple
pgvector